Removed old UI transform code, replaced with constraints.
Again build is broken, old SetTransform() commented out with new GetConstraints() equivalents being written below.

+++ Sources/Fonts/Text.hpp
@@ -251,7 +251,7 @@ private:
 
 	std::unique_ptr<Model> m_model;
 	uint32_t m_numberLines = 0;
-	Vector2f m_lastSize;
+	Vector2i m_lastSize;
 
 	float m_fontSize = 12;
 	std::string m_string;
